+++ ui/src/main/java/com/kotcrab/vis/ui/widget/VisTextArea.java
@@ -105,7 +105,7 @@ public class VisTextArea extends VisTextField {
 				int i = start;
 				for (; i < end; i++)
 					if (glyphPositions[i] > x) break;
-				if (glyphPositions[i] - x <= x - glyphPositions[i - 1]) {
+				if (i > 0 && glyphPositions[i] - x <= x - glyphPositions[i - 1]) {
 					return Math.min(i, text.length());
 				}
 				return Math.max(0, i - 1);